wpt_babel/content-security-policy/nonce-hiding/nonces.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<div id="log">
</div>
<script>
 "use strict";

function _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }
function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }
function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }
function _iterableToArrayLimit(r, l) { var t = null == r ? null : "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (null != t) { var e, n, i, u, a = [], f = !0, o = !1; try { if (i = (t = t.call(r)).next, 0 === l) { if (Object(t) !== t) return; f = !1; } else for (; !(f = (e = i.call(t)).done) && (a.push(e.value), a.length !== l); f = !0); } catch (r) { o = !0, n = r; } finally { try { if (!f && null != t.return && (u = t.return(), Object(u) !== u)) return; } finally { if (o) throw n; } } return a; } }
function _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }
var namespace_url = {
  "HTML": "http://www.w3.org/1999/xhtml",
  "SVG": "http://www.w3.org/2000/svg"
};
var test_cases = [["meh", "HTML"], ["div", "HTML"], ["script", "HTML"], ["meh", "SVG"], ["svg", "SVG"], ["script", "SVG"], ["style", "HTML"], ["link", "HTML"]];
test_cases.forEach(function (_ref) {
  var _ref2 = _slicedToArray(_ref, 2),
    localName = _ref2[0],
    namespace = _ref2[1];
  test(function (t) {
    var element = document.createElementNS(namespace_url[namespace], localName);
    t.add_cleanup(function () {
      return element.remove();
    });
    assert_equals(element.nonce, "", "Initial IDL attribute value");
    assert_equals(element.getAttribute("nonce"), null, "Initial content attribute");
    element.setAttribute("nonce", "x");
    assert_equals(element.nonce, "x", "IDL attribute is modified after content attribute set");
    assert_equals(element.getAttribute("nonce"), "x", "Content attribute is modified after content attribute set");
    document.body.appendChild(element);
    assert_equals(element.nonce, "x", "IDL attribute is unchanged after element insertion");
    assert_equals(element.getAttribute("nonce"), "", "Content attribute is changed after element insertion");
  }, "Basic nonce tests for ".concat(localName, " in ").concat(namespace, " namespace"));
  test(function (t) {
    var element = document.createElementNS(namespace_url[namespace], localName);
    t.add_cleanup(function () {
      return element.remove();
    });
    element.setAttribute("nonce", "x");
    assert_equals(element.nonce, "x", "IDL attribute is modified after content attribute set");
    element.removeAttribute("nonce");
    assert_equals(element.nonce, "", "IDL attribute is empty after content attribute removal");
  }, "Ensure that removal of content attribute does not affect IDL attribute for ".concat(localName, " in ").concat(namespace, " namespace"));
  test(function (t) {
    var element = document.createElementNS(namespace_url[namespace], localName);
    t.add_cleanup(function () {
      return element.remove();
    });
    assert_equals(element.nonce, "");
    assert_equals(element.getAttribute("nonce"), null);
    element.setAttribute("nonce", "");
    assert_equals(element.nonce, "");
    assert_equals(element.getAttribute("nonce"), "");
    document.body.appendChild(element);
    assert_equals(element.nonce, "");
    assert_equals(element.getAttribute("nonce"), "");
    element.removeAttribute("nonce");
    assert_equals(element.nonce, "");
    assert_equals(element.getAttribute("nonce"), null);
  }, "Test empty nonces for ".concat(localName, " in ").concat(namespace, " namespace"));
});
</script>
